requirements.txt
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django>=2.2.15
django-bootstrap4==1.1.1
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
zipp==1.2.0
